requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1

# ORM + база
sqlalchemy==2.0.36
psycopg2-binary==2.9.9

# Конфигурации и окружение
python-dotenv==1.0.1

# Аутентификация / безопасность
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.8.2
python-multipart==0.0.9

# Миграции
alembic==1.13.3

# Тестирование / HTTP-клиент
pytest==8.3.2
httpx==0.27.2

# (опционально, но полезно)
loguru==0.7.2
